refactor(api): fold guide-variants-ranking test into guides test file
